break out sigma.utils into module files
