+++ src/domain/classes/Camera.js
