+++ src/domain/classes/Configurable.js
