+++ src/domain/classes/Dispatcher.js
