+++ src/domain/classes/MouseCaptor.js
