+++ src/domain/classes/TouchCaptor.js
